requirements.txt
pymupdf>=1.26
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.0
pytest>=7.0
pytest-mock>=3.10